PT-2669: Make the patient record form easier to configure from other modules Moved the default RecordSection and RecordElement implementation as public SPIs
